Fix #1, the right function is "IsMouseButtonPressed()"
Signs are placed automatically when you keep pressed the left mouse button and move the mouse over the grid. To resolve this bug, I've changed the function that detects the user's input: "IsMouseButtonDown()" detects if a mouse button is being pressed, while "IsMouseButtonPressed()" detects if a mouse button has been pressed once; so I use this last function.